Add backend

+++ AdminASP/Controllers/CthdController.cs
@@ -12,56 +12,60 @@ namespace AdminASP.Controllers
 {
     public class CthdController : Controller
     {
-        public IActionResult GetAll()
+        public String GetAll()
         {
+            if (!(CheckPermission.CheckStaff(this))) { return ""; }
+
             CthdStoreContext modelStoreContext = HttpContext.RequestServices.GetService(typeof(CthdStoreContext)) as CthdStoreContext;
             List<BaseModel> baseModels = modelStoreContext.GetAll();
-            List<Cthd> thisModels = new List<Cthd>();
+            List<Cthd> outputs = new List<Cthd>();
             foreach (BaseModel baseModel in baseModels)
             {
-                thisModels.Add(baseModel as Cthd);
+                outputs.Add(baseModel as Cthd);
             }
-            ViewData["inputs"] = thisModels;
-            return View();
+
+            return JsonConvert.SerializeObject(outputs);
         }
 
-        public IActionResult Add(FormCthdAddInput input)
+        public String Add(FormCthdAddInput input)
         {
-            int result = 0;
-            List<String> resultValidate = input.GetValidate();
-            if (resultValidate.Count <= 0)
+            if (!(CheckPermission.CheckAdmin(this))) { return ""; }
+
+            int output = 0;
+            Cthd newCthd = null;
+            List<String> errors = input.GetValidate();
+            if (errors.Count <= 0)
             {
                 CthdStoreContext modelStoreContext = HttpContext.RequestServices.GetService(typeof(CthdStoreContext)) as CthdStoreContext;
-                int addResult = modelStoreContext.Add(new Cthd()
+                newCthd = new Cthd()
                 {
                     IdHoaDon = input.IdHoaDon,
                     IdSanPham = input.IdSanPham,
                     SoLuong = input.SoLuong,
                     DonGia = input.DonGia,
                     DiemTichLuy = input.DiemTichLuy
-                });
+                };
+                int addResult = modelStoreContext.Add(newCthd);
 
-                result = addResult;
+                output = addResult;
             }
-            ViewData["input"] = result;
-            ViewData["errors"] = resultValidate;
-            return View();
+
+            return JsonConvert.SerializeObject(new { output = output, errors = errors, newCthd = newCthd });
         }
 
-        public IActionResult Edit(FormCthdEditInput input)
+        public String Edit(FormCthdEditInput input)
         {
-            int result = 0;
-            List<String> resultValidate = input.GetValidate();
-            if (resultValidate.Count <= 0)
+            if (!(CheckPermission.CheckAdmin(this))) { return ""; }
+
+            int output = 0;
+            List<String> errors = input.GetValidate();
+            if (errors.Count <= 0)
             {
                 CthdStoreContext modelStoreContext = HttpContext.RequestServices.GetService(typeof(CthdStoreContext)) as CthdStoreContext;
                 Cthd oldCthd = new Cthd()
                 {
                     IdHoaDon = input.IdHoaDon,
                     IdSanPham = input.IdSanPham,
-                    SoLuong = -1,
-                    DonGia = -1,
-                    DiemTichLuy = -1
                 };
                 Cthd newCthd = new Cthd()
                 {
@@ -73,47 +77,70 @@ namespace AdminASP.Controllers
                 };
                 int editResult = modelStoreContext.Edit(oldCthd, newCthd);
 
-                result = editResult;
-                ViewData["newCthd"] = newCthd;
+                output = editResult;
             }
-            ViewData["input"] = result;
-            ViewData["errors"] = resultValidate;
-            return View();
+
+            return JsonConvert.SerializeObject(new { output = output, errors = errors });
         }
 
-        public IActionResult Delete(FormCthdDeleteInput input)
+        public String Delete(FormCthdDeleteInput input)
         {
-            int result = 0;
-            List<String> resultValidate = input.GetValidate();
-            if (resultValidate.Count <= 0)
+            if (!(CheckPermission.CheckAdmin(this))) { return ""; }
+
+            int output = 0;
+            List<String> errors = input.GetValidate();
+            if (errors.Count <= 0)
             {
                 CthdStoreContext modelStoreContext = HttpContext.RequestServices.GetService(typeof(CthdStoreContext)) as CthdStoreContext;
                 Cthd taiKhoan = new Cthd()
                 {
                     IdHoaDon = input.IdHoaDon,
                     IdSanPham = input.IdSanPham,
-                    SoLuong = -1,
-                    DonGia = -1,
-                    DiemTichLuy = -1
                 };
                 int deleteResult = modelStoreContext.Delete(taiKhoan);
 
-                result = deleteResult;
+                output = deleteResult;
             }
-            ViewData["input"] = result;
-            ViewData["errors"] = resultValidate;
-            return View();
+
+            return JsonConvert.SerializeObject(new { output = output, errors = errors });
         }
 
-        public IActionResult DeleteAll()
+        public String DeleteMarked(FormCthdDeleteMarked deleteInput)
         {
-            int result = 0;
+            if (!(CheckPermission.CheckAdmin(this))) { return ""; }
+
+            List<Cthd> inputs = JsonConvert.DeserializeObject<List<Cthd>>(deleteInput.JsonInput);
+
+            List<String> outputs = new List<String>();
+            if (inputs.Count > 0)
+            {
+                CthdStoreContext modelStoreContext = HttpContext.RequestServices.GetService(typeof(CthdStoreContext)) as CthdStoreContext;
+                foreach (Cthd input in inputs)
+                {
+                    Cthd cthd = new Cthd()
+                    {
+                        IdHoaDon = input.IdHoaDon,
+                        IdSanPham = input.IdSanPham
+                    };
+                    int deleteResult = modelStoreContext.Delete(cthd);
+                    outputs.Add(JsonConvert.SerializeObject(new { IdHoaDon = input.IdHoaDon, IdSanPham = input.IdSanPham, Result = deleteResult }));
+                }
+            }
+
+            return JsonConvert.SerializeObject(outputs);
+        }
+
+        public String DeleteAll()
+        {
+            if (!(CheckPermission.CheckAdmin(this))) { return ""; }
+
+            int output = 0;
             CthdStoreContext modelStoreContext = HttpContext.RequestServices.GetService(typeof(CthdStoreContext)) as CthdStoreContext;
             int deleteResult = modelStoreContext.DeleteAll();
 
-            result = deleteResult;
-            ViewData["input"] = result;
-            return View();
+            output = deleteResult;
+
+            return JsonConvert.SerializeObject(new { output = output });
         }
     }
 }
